Splitting the types definitions contained in action.types & generic.actions
