+++ front-react/src/redux/types/action.result.ts
@@ -1,0 +1,5 @@
+import ApplicationError from "./application.error.type";
+
+export interface ActionResult extends ApplicationError {
+  success: boolean;
+}
