+++ front-react/src/redux/types/thunk.result.ts
@@ -1,0 +1,4 @@
+import { ThunkAction } from "redux-thunk";
+import { RootState } from "../store/root.state";
+
+export type ThunkResult<R> = ThunkAction<R, RootState, undefined, any>;
